refactor face detection
