+++ src/routes/index.tsx
@@ -7,13 +7,12 @@ import { object, optional, string } from "valibot";
 import Header from "~/components/Header";
 import Icon from "~/components/Icon";
 import Settings from "~/components/Settings";
-import { CameraProvider } from "~/context/camera";
+import { FaceDetectProvider } from "~/context/faceDetect";
 import {
   IconColorsProvider,
   parseColors,
   useIconColors,
 } from "~/context/iconColors";
-import { IconTransformsProvider } from "~/context/iconTransforms";
 import { useQuery } from "~/lib/query";
 
 const querySchema = object({
@@ -102,37 +101,35 @@ export default function Home() {
         type="image/svg+xml"
         href={`/image?f=svg${serverIconParam ? `&p=${serverIconParam}` : ""}`}
       />
-      <CameraProvider>
-        <IconTransformsProvider>
-          <IconColorsProvider params={param()}>
-            <Show when={!isFull()} fallback={<IconWrapper />}>
-              <div class="grid grid-rows-[auto_1fr] w-full h-full prose prose-zinc max-w-unset! overflow-hidden">
-                <Header />
-                <main class="h-full w-full overflow-hidden">
-                  <Splitter.Root
-                    orientation={matches.md ? "horizontal" : "vertical"}
-                    size={[
-                      { id: "icon", size: 50, minSize: 20 },
-                      { id: "settings", size: 50, minSize: 20 },
-                    ]}
-                  >
-                    <Splitter.Panel id="icon">
-                      <IconWrapper />
-                    </Splitter.Panel>
-                    <Splitter.ResizeTrigger
-                      class="outline-none min-w-1 min-h-1 bg-zinc-200"
-                      id="icon:settings"
-                    />
-                    <Splitter.Panel id="settings">
-                      <Settings />
-                    </Splitter.Panel>
-                  </Splitter.Root>
-                </main>
-              </div>
-            </Show>
-          </IconColorsProvider>
-        </IconTransformsProvider>
-      </CameraProvider>
+      <FaceDetectProvider>
+        <IconColorsProvider params={param()}>
+          <Show when={!isFull()} fallback={<IconWrapper />}>
+            <div class="grid grid-rows-[auto_1fr] w-full h-full prose prose-zinc max-w-unset! overflow-hidden">
+              <Header />
+              <main class="h-full w-full overflow-hidden">
+                <Splitter.Root
+                  orientation={matches.md ? "horizontal" : "vertical"}
+                  size={[
+                    { id: "icon", size: 50, minSize: 20 },
+                    { id: "settings", size: 50, minSize: 20 },
+                  ]}
+                >
+                  <Splitter.Panel id="icon">
+                    <IconWrapper />
+                  </Splitter.Panel>
+                  <Splitter.ResizeTrigger
+                    class="outline-none min-w-1 min-h-1 bg-zinc-200"
+                    id="icon:settings"
+                  />
+                  <Splitter.Panel id="settings">
+                    <Settings />
+                  </Splitter.Panel>
+                </Splitter.Root>
+              </main>
+            </div>
+          </Show>
+        </IconColorsProvider>
+      </FaceDetectProvider>
     </>
   );
 }
